P2P: Fix P2P Group Formation timeout registration (use parent iface)
Commit 361cdf34009c7407ddf22dee6f5e6ce82e81bcae extended the use of
group formation timeout for the way handshake, but the registration was
done on the group_interface while the cancellation was done on the
parent interface. Fix the registration to set the eloop timeout on
parent to address potential issues when using a separate group
interface.

Signed-hostap: Nirav Shah <nirav.j2.shah@intel.com>